+++ app/forms/waste_exemptions_engine/operator_address_lookup_form.rb
@@ -2,24 +2,7 @@
 
 module WasteExemptionsEngine
   class OperatorAddressLookupForm < AddressLookupForm
-    include CanNavigateFlexibly
+    include OperatorAddressForm
 
-    attr_accessor :business_type
-
-    def initialize(enrollment)
-      super
-      # We only use this for the correct microcopy
-      self.business_type = @enrollment.business_type
-    end
-
-    private
-
-    def existing_address
-      @enrollment.operator_address
-    end
-
-    def address_type
-      Address.address_types[:operator]
-    end
   end
 end
